cloud_function/requirements.txt
functions-framework
google-generativeai
